test: replace tap with jest
